backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
librosa>=0.9.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.20.0
python-multipart>=0.0.5
noisereduce>=2.0.0
scipy>=1.7.0
scikit-learn>=1.0.0